[MGCB Editor] Update registration handling files
- Some code formatting/improved readability for Mac and Linux FileAssociation files
- Removed all messages about association except the final result message
- Failed registration now always gives a stacktrace
- Unassociation is now always called before Association

+++ Tools/MonoGame.Content.Builder.Editor/Platform/Windows/FileAssociation.Windows.cs
@@ -70,72 +70,59 @@ namespace MonoGame.Tools.Pipeline
             var command = $"\"{startInfo.FileName}\" {startInfo.Arguments}";
             var iconPath = $"{assembly.Location},0";
 
-            SetWindowsAssociation(extension, progId, fileTypeDescription, iconPath, verb, commandText, command);
-            RefreshEnvironment();
-
             var link = (IShellLink)new ShellLink();
             link.SetPath(Path.Combine(Path.GetDirectoryName(typeof(FileAssociation).Assembly.Location), "mgcb-editor-wpf.exe"));
             ((IPersistFile)link).Save(startMenuLink, false);
+
+            SetWindowsAssociation(extension, progId, fileTypeDescription, iconPath, verb, commandText, command);
+            RefreshEnvironment();
         }
 
         public static void Unassociate()
         {
+            if (File.Exists(startMenuLink))
+                File.Delete(startMenuLink);
+
             UnsetWindowsAssociation(extension, progId);
             RefreshEnvironment();
-
-            File.Delete(startMenuLink);
         }
 
-        private static bool SetWindowsAssociation(string extension, string progId, string fileTypeDescription, string icon, string verb, string commandText, string command)
+        private static void SetWindowsAssociation(string extension, string progId, string fileTypeDescription, string icon, string verb, string commandText, string command)
         {
-            var madeChanges = false;
-            Console.WriteLine($"Associating MGCB Editor with .{extension} extension in Windows...");
-            madeChanges |= SetRegistryKey($@"Software\Classes\.{extension}", null, progId);
-            madeChanges |= SetRegistryKey($@"Software\Classes\{progId}", null, fileTypeDescription);
-            madeChanges |= SetRegistryKey($@"Software\Classes\{progId}\DefaultIcon", null, icon);
-            madeChanges |= SetRegistryKey($@"Software\Classes\{progId}\shell", null, verb);
-            madeChanges |= SetRegistryKey($@"Software\Classes\{progId}\shell\{verb}", null, commandText);
-            madeChanges |= SetRegistryKey($@"Software\Classes\{progId}\shell\{verb}\command", null, command);
-            Console.WriteLine(madeChanges ? $"Association complete!" : "Already associated.");
-            return madeChanges;
+            SetRegistryKey($@"Software\Classes\.{extension}", null, progId);
+            SetRegistryKey($@"Software\Classes\{progId}", null, fileTypeDescription);
+            SetRegistryKey($@"Software\Classes\{progId}\DefaultIcon", null, icon);
+            SetRegistryKey($@"Software\Classes\{progId}\shell", null, verb);
+            SetRegistryKey($@"Software\Classes\{progId}\shell\{verb}", null, commandText);
+            SetRegistryKey($@"Software\Classes\{progId}\shell\{verb}\command", null, command);
         }
 
-        private static bool UnsetWindowsAssociation(string extension, string progId)
+        private static void UnsetWindowsAssociation(string extension, string progId)
         {
-            var madeChanges = false;
-            Console.WriteLine($"Unassociating MGCB Editor with .{extension} extension in Windows...");
-            madeChanges |= DeleteRegistryKeyTree($@"Software\Classes\.{extension}");
-            madeChanges |= DeleteRegistryKeyTree($@"Software\Classes\{progId}");
-            Console.WriteLine(madeChanges ? $"Unassociation complete!" : "No association found.");
-            return madeChanges;
+            DeleteRegistryKeyTree($@"Software\Classes\.{extension}");
+            DeleteRegistryKeyTree($@"Software\Classes\{progId}");
         }
 
-        private static bool SetRegistryKey<T>(string keyPath, string name, T value, RegistryValueKind valueKind = RegistryValueKind.String)
+        private static void SetRegistryKey<T>(string keyPath, string name, T value, RegistryValueKind valueKind = RegistryValueKind.String)
         {
             using (var key = Registry.CurrentUser.CreateSubKey(keyPath))
             {
                 if (!value.Equals(key.GetValue(name)))
                 {
                     key.SetValue(name, value, valueKind);
-                    return true;
                 }
             }
-
-            return false;
         }
 
-        private static bool DeleteRegistryKeyTree(string keyPath)
+        private static void DeleteRegistryKeyTree(string keyPath)
         {
             using (var key = Registry.CurrentUser.OpenSubKey(keyPath))
             {
                 if (key != null)
                 {
                     Registry.CurrentUser.DeleteSubKeyTree(keyPath);
-                    return true;
                 }
             }
-
-            return false;
         }
 
         private static void RefreshEnvironment()
